dfuse: Only wait for poll timeout if in dfuDNBUSY state
The DFU 1.1 specification only mentions "bwPollTimeout elapsed" for the
dfuBUSY state (Appendix A.2.5).

In dfuDNLOAD-IDLE state (Appendix A.2.6) it seems not to be relevant.
We should therefore be able to ignore the poll timeout and not wait in
this state.

Note we already do this for normal DFU (non-DfuSE) devices.

This can speed up operation considerably on DfuSE devices that report a
non-null bwPollTimeout in this state.

https://sourceforge.net/p/dfu-util/tickets/193/

Signed-off-by: Tormod Volden <debian.tormod@gmail.com>