dfuse: Retry get_status if pipe error on download
If something goes wrong we cannot keep trying forever, so limit it to 3
retries for now.

This is similar to what is already done for special commands, originally
introduced because some bootloaders report too short poll timeouts.
Update the comment there to reflect the utility also for a "fast" mode.

https://sourceforge.net/p/dfu-util/tickets/193/

This might also address pipe errors reported other places, like in

https://sourceforge.net/p/dfu-util/tickets/46/

Signed-off-by: Tormod Volden <debian.tormod@gmail.com>